add chat content
